rename `account_builder.rs` to `wasm_account_builder.rs`
