+++ bindings/wasm/src/account/mod.rs
@@ -1,9 +1,9 @@
 // Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod account_builder;
 mod update;
 mod wasm_account;
+mod wasm_account_builder;
 mod wasm_auto_save;
 mod wasm_identity_setup;
 mod wasm_method_relationship;
